making msgdata write and read consistent
