+++ templates/stuff/stuff.yaml
@@ -1,0 +1,11 @@
+'1':
+- 1
+- 2
+- 3
+crop_b: 0
+crop_l: 0
+crop_r: 0
+crop_t: 0
+scale: []
+scale_loc_horz: 0
+scale_loc_vert: 0
